\name{LARkmeans-package}
\alias{LARkmeans-package}
\alias{LARkmeans}
\docType{package}
\title{
\packageTitle{LARkmeans}
}
\description{
\packageDescription{LARkmeans}
}
\details{
\packageDESCRIPTION{LARkmeans}
\packageIndices{LARkmeans}
}
\author{
\packageAuthor{LARkmeans}

Maintainer: \packageMaintainer{LARkmeans}
}
\references{
Forgy, E. W. (1965) Cluster analysis of multivariate data: efficiency vs interpretability of classifications. Biometrics \strong{21}, 768--769.

Hamerly, G.; Elkan, C. (2002) \href{http://people.csail.mit.edu/tieu/notebook/kmeans/15_p600-hamerly.pdf}{Alternatives to the k-means algorithm that find better clusterings} (PDF). Proceedings of the eleventh international conference on Information and knowledge management (CIKM).

Lloyd, S. P. (1957, 1982) Least squares quantization in PCM. Technical Note, Bell Laboratories. Published in 1982 in IEEE Transactions on Information Theory \strong{28}, 128--137.
}
\keyword{ package }
\keyword{ cluster }